move fx output query
