+++ ipfx/lims_queries.py
@@ -1,4 +1,5 @@
 import os
+import glob
 import logging
 import pg8000
 from ipfx.py2to3 import to_str
@@ -201,3 +202,47 @@ def project_specimen_ids(project, passed_only=True):
     results = query(SQL)
     sp_ids = [d["id"] for d in results]
     return sp_ids
+
+
+def get_fx_output_json(specimen_id):
+    """
+    Find in LIMS the full path to the json output of the feature extraction module
+    If more than one file exists, then chose the latest version
+
+    Parameters
+    ----------
+    specimen_id
+
+    Returns
+    -------
+    file_name: string
+    """
+    NO_SPECIMEN = "No_specimen_in_LIMS"
+    NO_OUTPUT_FILE = "No_feature_extraction_output"
+    
+    sql = """
+    select err.storage_directory, err.id
+    from specimens sp
+    join ephys_roi_results err on err.id = sp.ephys_roi_result_id
+    where sp.id = %d
+    """ % specimen_id
+
+    res = query(sql)
+    if res:
+        err_dir = res[0]["storage_directory"]
+
+        file_list = glob.glob(os.path.join(err_dir, '*EPHYS_FEATURE_EXTRACTION_*_output.json'))
+        if file_list:
+            latest_file = max(file_list, key=os.path.getctime)   # get the most recent file
+            return latest_file
+        else:
+            return NO_OUTPUT_FILE
+    else:
+        return NO_SPECIMEN
+
+
+def fix_network_path(lims_path):
+    # Need to have double slash for network drive
+    if not lims_path.startswith('//'):
+        lims_path = '/' + lims_path
+    return str(Path(lims_path))
